claimDetailsController created, extracted from citizenDetailsController
