+++ src/main/routes/features/response/citizenDetails/citizenDetailsController.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 import { Validator } from 'class-validator';
 import config from 'config';
 import { Form } from '../../../../common/form/models/form';
-import {CLAIM_DETAILS_URL, CITIZEN_DETAILS_URL } from '../../../urls';
+import { CITIZEN_DETAILS_URL } from '../../../urls';
 import { CivilServiceClient } from '../../../../app/client/civilServiceClient';
 import {Claim} from '../../../../common/models/claim';
 import { CitizenAddress } from '../../../../common/form/models/citizenAddress';
@@ -15,18 +15,6 @@ const civilServiceClient: CivilServiceClient = new CivilServiceClient(civilServi
 let citizenFullName: object;
 
 let claim: Claim = new Claim();
-
-function renderPage(res: express.Response, claimDetails: Claim): void {
-  res.render('features/response/claim-details', {
-    claim: claimDetails,
-  });
-}
-
-// -- GET Claim Details
-router.get(CLAIM_DETAILS_URL, async (req: express.Request, res: express.Response) => {
-  claim = await civilServiceClient.retrieveClaimDetails('1643033241924739');
-  renderPage(res, claim);
-});
 
 // -- GET Citizen Details
 router.get(CITIZEN_DETAILS_URL, async (req: express.Request, res: express.Response) => {
